+++ src/main/java/many/studio/web_backend/entity/Anamnese.java
@@ -8,20 +8,20 @@ import java.time.LocalDateTime;
 public class Anamnese {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "informacao")
     private String informacao;
 
     public Anamnese() {}
 
-    public Anamnese(Integer id, String informacao) {
+    public Anamnese(Long id, String informacao) {
         this.id = id;
         this.informacao = informacao;
     }
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getInformacao() { return informacao; }
     public void setInformacao(String informacao) { this.informacao = informacao; }
 }
