+++ src/main/java/many/studio/web_backend/entity/TipoPagamento.java
@@ -8,20 +8,20 @@ import jakarta.persistence.Id;
 public class TipoPagamento {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "tipo")
     private String tipo;
 
     public TipoPagamento() {}
 
-    public TipoPagamento(Integer id, String tipo) {
+    public TipoPagamento(Long id, String tipo) {
         this.id = id;
         this.tipo = tipo;
     }
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getTipo() { return tipo; }
     public void setTipo(String tipo) { this.tipo = tipo; }
 }
