+++ src/main/java/many/studio/web_backend/repository/UsuarioRepository.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
     Optional<Usuario> findByEmail(String email);
     boolean existsByEmail(String email);
-    boolean existsById(Integer id);
+    boolean existsById(Long id);
 
 }
